+++ utils/__init__.py
@@ -76,6 +76,7 @@ from .creator_metrics import (
     estimate_monthly_revenue_v4,
     format_channel_age,
     get_activity_badge,
+    get_country_name,
     get_grade_info,
     get_growth_signal,
     get_language_emoji,
